+++ examples/data-sources/meraki_camera_wireless_profile/data-source.tf
@@ -1,0 +1,4 @@
+data "meraki_camera_wireless_profile" "example" {
+  id         = "12345678"
+  network_id = "L_123456"
+}