+++ examples/data-sources/meraki_camera_wireless_profiles/data-source.tf
@@ -1,0 +1,3 @@
+data "meraki_camera_wireless_profiles" "example" {
+  network_id = "L_123456"
+}