+++ examples/resources/meraki_camera_wireless_profile/import.sh
@@ -1,0 +1,1 @@
+terraform import meraki_camera_wireless_profile.example "<network_id>,<id>"