+++ internal/provider/resource_meraki_camera_wireless_profile_test.go
@@ -1,0 +1,136 @@
+// Copyright © 2024 Cisco Systems, Inc. and its affiliates.
+// All rights reserved.
+//
+// Licensed under the Mozilla Public License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://mozilla.org/MPL/2.0/
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// SPDX-License-Identifier: MPL-2.0
+
+package provider
+
+// Section below is generated&owned by "gen/generator.go". //template:begin imports
+import (
+	"fmt"
+	"os"
+	"testing"
+
+	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
+	"github.com/hashicorp/terraform-plugin-testing/terraform"
+)
+
+// End of section. //template:end imports
+
+// Section below is generated&owned by "gen/generator.go". //template:begin testAcc
+
+func TestAccMerakiCameraWirelessProfile(t *testing.T) {
+	if os.Getenv("TF_VAR_test_org") == "" || os.Getenv("TF_VAR_test_network") == "" {
+		t.Skip("skipping test, set environment variable TF_VAR_test_org and TF_VAR_test_network")
+	}
+	var checks []resource.TestCheckFunc
+	checks = append(checks, resource.TestCheckResourceAttr("meraki_camera_wireless_profile.test", "name", "wireless profile A"))
+	checks = append(checks, resource.TestCheckResourceAttr("meraki_camera_wireless_profile.test", "identity_password", "password123"))
+	checks = append(checks, resource.TestCheckResourceAttr("meraki_camera_wireless_profile.test", "identity_username", "identityname"))
+	checks = append(checks, resource.TestCheckResourceAttr("meraki_camera_wireless_profile.test", "ssid_auth_mode", "8021x-radius"))
+	checks = append(checks, resource.TestCheckResourceAttr("meraki_camera_wireless_profile.test", "ssid_encryption_mode", "wpa-eap"))
+	checks = append(checks, resource.TestCheckResourceAttr("meraki_camera_wireless_profile.test", "ssid_name", "ssid test"))
+
+	var steps []resource.TestStep
+	if os.Getenv("SKIP_MINIMUM_TEST") == "" {
+		steps = append(steps, resource.TestStep{
+			Config: testAccMerakiCameraWirelessProfilePrerequisitesConfig + testAccMerakiCameraWirelessProfileConfig_minimum(),
+		})
+	}
+	steps = append(steps, resource.TestStep{
+		Config: testAccMerakiCameraWirelessProfilePrerequisitesConfig + testAccMerakiCameraWirelessProfileConfig_all(),
+		Check:  resource.ComposeTestCheckFunc(checks...),
+	})
+	steps = append(steps, resource.TestStep{
+		ResourceName:            "meraki_camera_wireless_profile.test",
+		ImportState:             true,
+		ImportStateVerify:       true,
+		ImportStateIdFunc:       merakiCameraWirelessProfileImportStateIdFunc("meraki_camera_wireless_profile.test"),
+		ImportStateVerifyIgnore: []string{},
+		Check:                   resource.ComposeTestCheckFunc(checks...),
+	})
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:                 func() { testAccPreCheck(t) },
+		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+		Steps:                    steps,
+	})
+}
+
+// End of section. //template:end testAcc
+
+// Section below is generated&owned by "gen/generator.go". //template:begin importStateIdFunc
+
+func merakiCameraWirelessProfileImportStateIdFunc(resourceName string) resource.ImportStateIdFunc {
+	return func(s *terraform.State) (string, error) {
+		primary := s.RootModule().Resources[resourceName].Primary
+		NetworkId := primary.Attributes["network_id"]
+		id := primary.Attributes["id"]
+
+		return fmt.Sprintf("%s,%s", NetworkId, id), nil
+	}
+}
+
+// End of section. //template:end importStateIdFunc
+
+// Section below is generated&owned by "gen/generator.go". //template:begin testPrerequisites
+
+const testAccMerakiCameraWirelessProfilePrerequisitesConfig = `
+variable "test_org" {}
+variable "test_network" {}
+data "meraki_organization" "test" {
+  name = var.test_org
+}
+resource "meraki_network" "test" {
+  organization_id = data.meraki_organization.test.id
+  name            = var.test_network
+  product_types   = ["switch", "wireless", "appliance", "sensor", "camera"]
+}
+
+`
+
+// End of section. //template:end testPrerequisites
+
+// Section below is generated&owned by "gen/generator.go". //template:begin testAccConfigMinimal
+
+func testAccMerakiCameraWirelessProfileConfig_minimum() string {
+	config := `resource "meraki_camera_wireless_profile" "test" {` + "\n"
+	config += `  network_id = meraki_network.test.id` + "\n"
+	config += `  name = "wireless profile A"` + "\n"
+	config += `  ssid_auth_mode = "psk"` + "\n"
+	config += `  ssid_name = "ssid test"` + "\n"
+	config += `  ssid_psk = "password123"` + "\n"
+	config += `}` + "\n"
+	return config
+}
+
+// End of section. //template:end testAccConfigMinimal
+
+// Section below is generated&owned by "gen/generator.go". //template:begin testAccConfigAll
+
+func testAccMerakiCameraWirelessProfileConfig_all() string {
+	config := `resource "meraki_camera_wireless_profile" "test" {` + "\n"
+	config += `  network_id = meraki_network.test.id` + "\n"
+	config += `  name = "wireless profile A"` + "\n"
+	config += `  identity_password = "password123"` + "\n"
+	config += `  identity_username = "identityname"` + "\n"
+	config += `  ssid_auth_mode = "8021x-radius"` + "\n"
+	config += `  ssid_encryption_mode = "wpa-eap"` + "\n"
+	config += `  ssid_name = "ssid test"` + "\n"
+	config += `}` + "\n"
+	return config
+}
+
+// End of section. //template:end testAccConfigAll